requirements.txt
django-taggit==0.9.3
django-mptt==0.5.5
-e git://github.com/mauler/django-extensions.git#egg=django_extensions-dev
